Update banners conformance
