+++ Source/ChartboostAdapterAd.swift
@@ -7,8 +7,8 @@ import ChartboostMediationSDK
 import ChartboostSDK
 import Foundation
 
-/// Chartboost Mediation Chartboost adapter ad.
-final class ChartboostAdapterAd: NSObject, PartnerAd {
+/// Base class for Chartboost Mediation Chartboost adapter ads.
+final class ChartboostAdapterAd: NSObject {
     
     /// The partner adapter that created this ad.
     let adapter: PartnerAdapter
@@ -24,85 +24,25 @@ final class ChartboostAdapterAd: NSObject, PartnerAd {
     /// It should be the one provided on ``PartnerAdapter/makeBannerAd(request:delegate:)`` or ``PartnerAdapter/makeFullscreenAd(request:delegate:)``.
     weak var delegate: PartnerAdDelegate?
     
-    /// The partner ad view to display inline. E.g. a banner view.
-    /// Should be nil for full-screen ads.
-    var inlineView: UIView? { chartboostAd as? UIView }
-    
-    /// The loaded partner ad banner size.
-    /// Should be `nil` for full-screen ads.
-    var bannerSize: PartnerBannerSize?
-
-    /// The Chartboost SDK ad.
-    private let chartboostAd: CHBAd
-        
     /// The completion for the ongoing load operation.
     private var loadCompletion: ((Result<PartnerDetails, Error>) -> Void)?
     
     /// The completion for the ongoing show operation.
     private var showCompletion: ((Result<PartnerDetails, Error>) -> Void)?
     
+    /// The mediation object used to created Chartboost ads.
+    var mediation: CHBMediation {
+        .init(
+            name: "Chartboost",
+            libraryVersion: ChartboostMediation.sdkVersion,
+            adapterVersion: adapter.adapterVersion
+        )
+    }
+
     init(adapter: PartnerAdapter, request: PartnerAdLoadRequest, delegate: PartnerAdDelegate) throws {
         self.request = request
         self.delegate = delegate
         self.adapter = adapter
-        self.chartboostAd = try Self.makeChartboostAd(adapter: adapter, request: request)
-        let chartboostBannerSize = (self.chartboostAd as? CHBBanner)?.size
-        self.bannerSize = chartboostBannerSize.map { PartnerBannerSize(size: $0, type: .fixed) }
-    }
-    
-    /// Loads an ad.
-    /// - parameter viewController: The view controller on which the ad will be presented on. Needed on load for some banners.
-    /// - parameter completion: Closure to be performed once the ad has been loaded.
-    func load(with viewController: UIViewController?, completion: @escaping (Result<PartnerDetails, Error>) -> Void) {
-        log(.loadStarted)
-        
-        // Save load completion to execute later on didCacheAd.
-        // Banners are expected to show immediately after loading, so we call show() on load for banner ads only
-        if request.format == PartnerAdFormats.banner {
-            // Banners require a view controller on load to be able to show
-            guard let viewController = viewController else {
-                let error = error(.showFailureViewControllerNotFound)
-                log(.loadFailed(error))
-                completion(.failure(error))
-                return
-            }
-            // Banners show on load
-            loadCompletion = { [weak self] result in
-                if case .success = result {
-                    self?.chartboostAd.show(from: viewController)
-                }
-                completion(result)
-            }
-        } else {
-            // Interstitial and rewarded ads don't do anything extra
-            loadCompletion = completion
-        }
-        
-        // Set delegate
-        chartboostAd.delegate = self
-        
-        // Load the ad
-        if let adm = request.adm {
-            // Programmatic load
-            chartboostAd.cache(bidResponse: adm)
-        } else {
-            // Non-programmatic load
-            chartboostAd.cache()
-        }
-    }
-    
-    /// Shows a loaded ad.
-    /// Chartboost Mediation SDK will always call this method from the main thread.
-    /// - parameter viewController: The view controller on which the ad will be presented on.
-    /// - parameter completion: Closure to be performed once the ad has been shown.
-    func show(with viewController: UIViewController, completion: @escaping (Result<PartnerDetails, Error>) -> Void) {
-        log(.showStarted)
-        
-        // Save show completion to execute later on didShowAd
-        showCompletion = completion
-        
-        // Show the ad
-        chartboostAd.show(from: viewController)
     }
 }
 
@@ -162,72 +102,5 @@ extension ChartboostAdapterAd: CHBInterstitialDelegate, CHBRewardedDelegate, CHB
     
     func didFinishHandlingClick(_ event: CHBClickEvent, error: ClickError?) {
         log(.delegateCallIgnored)
-    }
-}
-
-private extension ChartboostAdapterAd {
-    
-    static func makeChartboostAd(adapter: PartnerAdapter, request: PartnerAdLoadRequest) throws -> CHBAd {
-        let mediation = CHBMediation(
-            name: "Chartboost",
-            libraryVersion: ChartboostMediation.sdkVersion,
-            adapterVersion: adapter.adapterVersion
-        )
-        switch request.format {
-        case PartnerAdFormats.interstitial:
-            return CHBInterstitial(
-                location: request.partnerPlacement,
-                mediation: mediation,
-                delegate: nil
-            )
-        case PartnerAdFormats.rewarded:
-            return CHBRewarded(
-                location: request.partnerPlacement,
-                mediation: mediation,
-                delegate: nil
-            )
-        case PartnerAdFormats.banner:
-            return try banner(adapter: adapter, request: request, mediation: mediation)
-        default:
-            throw adapter.error(.loadFailureUnsupportedAdFormat)
-        }
-    }
-}
-
-// MARK: - Helpers
-extension ChartboostAdapterAd {
-    private static func banner(
-        adapter: PartnerAdapter,
-        request: PartnerAdLoadRequest,
-        mediation: CHBMediation
-    ) throws -> CHBBanner {
-        // Fail if we cannot fit a fixed size banner in the requested size.
-        guard let size = fixedBannerSize(for: request.bannerSize) else {
-            throw adapter.error(.loadFailureInvalidBannerSize)
-        }
-
-        return CHBBanner(
-            size: size,
-            location: request.partnerPlacement,
-            mediation: mediation,
-            delegate: nil
-        )
-    }
-
-    private static func fixedBannerSize(for requestedSize: BannerSize?) -> CGSize? {
-        guard let requestedSize else {
-            return IABStandardAdSize
-        }
-        let sizes = [IABLeaderboardAdSize, IABMediumAdSize, IABStandardAdSize]
-        // Find the largest size that can fit in the requested size.
-        for size in sizes {
-            // If height is 0, the pub has requested an ad of any height, so only the width matters.
-            if requestedSize.size.width >= size.width &&
-                (size.height == 0 || requestedSize.size.height >= size.height) {
-                return size
-            }
-        }
-        // The requested size cannot fit any fixed size banners.
-        return nil
     }
 }
